+++ src/pages/kitchensink/Kitchensink.js
@@ -1,4 +1,5 @@
 import React from "react";
+import {Widget} from "../../shared/components";
 import DropDownExample from "./DropDownExample";
 import {withTranslation} from "react-i18next";
 import ButtonExample from "./ButtonExample";
@@ -8,20 +9,31 @@ import ListExample from "./ListExample";
 import BackendApi from "./BackendApi";
 
 class Kitchensink extends React.Component {
+    constructor(props) {
+        super(props);
+        this.state = {
+            show: false
+        }
+    }
 
-    test() {
-        alert("Test");
+    handleShow = () => {
+        this.setState({ show: true });
+    }
+
+    handleClose = () => {
+        this.setState({ show: false });
     }
 
     render() {
         return <div>
             <h1> { this.props.t('app.kitchensink.title') } </h1>
-            <ButtonExample handler={ this.test }/>
+            <ButtonExample handler={ this.handleShow }/>
             <DropDownExample/>
-            <ExpandableSectionExample handler={ this.test }/>
+            <ExpandableSectionExample handler={ this.handleShow }/>
             <TwoWayDatabingExample/>
-            <ListExample/>
+            <ListExample handler={ this.handleShow }/>
             <BackendApi/>
+            <Widget.Modal title={ this.props.t('app.kitchensink.modal.title') } closeFunction={ this.handleClose } message={ this.props.t('app.kitchensink.modal.message')} show={ this.state.show }/>
         </div>
     }
 }