+++ src/shared/components/index.js
@@ -4,6 +4,7 @@ import {APPEARANCE, NB_ITEM_TYPE, SIZE} from "./constants";
 import {WDropdown} from "./wdg.dropdown";
 import {WExpandableBody, WExpandableSection} from "./wdg.expandable-tab";
 import {WList} from "./wdg.list";
+import {WModal} from "./wdg.modal";
 
 export const Widget = {
 
@@ -15,6 +16,8 @@ export const Widget = {
     ExpandableSection: WExpandableSection,
     ExpandableTabBody: WExpandableBody,
     List: WList,
+
+    Modal: WModal,
 
     BUTTON_TYPE: APPEARANCE,
     SIZE: SIZE,